--- src/main/java/kitchenpos/external/domain/KitchenridersClient.java
@@ -1,8 +1,0 @@
-package kitchenpos.external.domain;
-
-import java.math.BigDecimal;
-import java.util.UUID;
-
-public interface KitchenridersClient {
-    void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
-}
