--- src/main/java/kitchenpos/external/infra/DefaultKitchenridersClient.java
@@ -1,14 +1,0 @@
-package kitchenpos.external.infra;
-
-import kitchenpos.external.domain.KitchenridersClient;
-import org.springframework.stereotype.Component;
-
-import java.math.BigDecimal;
-import java.util.UUID;
-
-@Component
-public class DefaultKitchenridersClient implements KitchenridersClient {
-    @Override
-    public void requestDelivery(final UUID orderId, final BigDecimal amount, final String deliveryAddress) {
-    }
-}
