+++ src/main/java/kitchenpos/order/common/model/OrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.common.model;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
+}
