+++ src/main/java/kitchenpos/order/common/model/OrderType.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.common.model;
+
+public enum OrderType {
+    DELIVERY, TAKEOUT, EAT_IN
+}
