+++ src/main/java/kitchenpos/order/deliveryorder/application/DeliveryOrderService.java
@@ -1,0 +1,157 @@
+package kitchenpos.order.deliveryorder.application;
+
+import kitchenpos.menu.domain.model.Menu;
+import kitchenpos.menu.domain.repository.MenuRepository;
+import kitchenpos.order.common.model.OrderLineItem;
+import kitchenpos.order.common.model.OrderStatus;
+import kitchenpos.order.common.model.OrderType;
+import kitchenpos.order.deliveryorder.domain.model.DeliveryOrder;
+import kitchenpos.order.deliveryorder.domain.repository.DeliveryOrderRepository;
+import kitchenpos.order.deliveryorder.external.domain.KitchenridersClient;
+import kitchenpos.order.eatinorder.domain.repository.OrderTableRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+import java.util.*;
+
+@Service
+public class DeliveryOrderService {
+
+    private final DeliveryOrderRepository deliveryOrderRepository;
+    private final MenuRepository menuRepository;
+    private final OrderTableRepository orderTableRepository;
+    private final KitchenridersClient kitchenridersClient;
+
+    public DeliveryOrderService(DeliveryOrderRepository deliveryOrderRepository, MenuRepository menuRepository, OrderTableRepository orderTableRepository, KitchenridersClient kitchenridersClient) {
+        this.deliveryOrderRepository = deliveryOrderRepository;
+        this.menuRepository = menuRepository;
+        this.orderTableRepository = orderTableRepository;
+        this.kitchenridersClient = kitchenridersClient;
+    }
+
+    @Transactional
+    public DeliveryOrder create(final DeliveryOrder request) {
+        final OrderType type = request.getType();
+        if (Objects.isNull(type)) {
+            throw new IllegalArgumentException();
+        }
+        final List<OrderLineItem> orderLineItemRequests = request.getOrderLineItems();
+        if (Objects.isNull(orderLineItemRequests) || orderLineItemRequests.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+        final List<Menu> menus = menuRepository.findAllByIdIn(
+                orderLineItemRequests.stream()
+                        .map(OrderLineItem::getMenuId)
+                        .toList()
+        );
+        if (menus.size() != orderLineItemRequests.size()) {
+            throw new IllegalArgumentException();
+        }
+        final List<OrderLineItem> orderLineItems = new ArrayList<>();
+        for (final OrderLineItem orderLineItemRequest : orderLineItemRequests) {
+            final long quantity = orderLineItemRequest.getQuantity();
+            final Menu menu = menuRepository.findById(orderLineItemRequest.getMenuId())
+                    .orElseThrow(NoSuchElementException::new);
+            if (!menu.isDisplayed()) {
+                throw new IllegalStateException();
+            }
+            if (menu.getPrice().compareTo(orderLineItemRequest.getPrice()) != 0) {
+                throw new IllegalArgumentException();
+            }
+            final OrderLineItem orderLineItem = new OrderLineItem();
+            orderLineItem.setMenu(menu);
+            orderLineItem.setQuantity(quantity);
+            orderLineItems.add(orderLineItem);
+        }
+
+        final String deliveryAddress = request.getDeliveryAddress();
+        if (Objects.isNull(deliveryAddress) || deliveryAddress.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+
+        DeliveryOrder order = new DeliveryOrder();
+        order.setId(UUID.randomUUID());
+        order.setStatus(OrderStatus.WAITING);
+        order.setOrderDateTime(LocalDateTime.now());
+        order.setOrderLineItems(orderLineItems);
+        order.setDeliveryAddress(deliveryAddress);
+
+        return deliveryOrderRepository.save(order);
+    }
+
+    @Transactional
+    public DeliveryOrder accept(final UUID orderId) {
+        final DeliveryOrder order = deliveryOrderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        if (order.getStatus() != OrderStatus.WAITING) {
+            throw new IllegalStateException();
+        }
+
+        BigDecimal sum = BigDecimal.ZERO;
+        for (final OrderLineItem orderLineItem : order.getOrderLineItems()) {
+            sum = orderLineItem.getMenu()
+                    .getPrice()
+                    .multiply(BigDecimal.valueOf(orderLineItem.getQuantity()));
+        }
+        kitchenridersClient.requestDelivery(orderId, sum, order.getDeliveryAddress());
+
+        order.setStatus(OrderStatus.ACCEPTED);
+        return order;
+    }
+
+    @Transactional
+    public DeliveryOrder serve(final UUID orderId) {
+        final DeliveryOrder order = deliveryOrderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        if (order.getStatus() != OrderStatus.ACCEPTED) {
+            throw new IllegalStateException();
+        }
+        order.setStatus(OrderStatus.SERVED);
+        return order;
+    }
+
+    @Transactional
+    public DeliveryOrder startDelivery(final UUID orderId) {
+        final DeliveryOrder order = deliveryOrderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        if (order.getStatus() != OrderStatus.SERVED) {
+            throw new IllegalStateException();
+        }
+        order.setStatus(OrderStatus.DELIVERING);
+        return order;
+    }
+
+    @Transactional
+    public DeliveryOrder completeDelivery(final UUID orderId) {
+        final DeliveryOrder order = deliveryOrderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        if (order.getStatus() != OrderStatus.DELIVERING) {
+            throw new IllegalStateException();
+        }
+        order.setStatus(OrderStatus.DELIVERED);
+        return order;
+    }
+
+    @Transactional
+    public DeliveryOrder complete(final UUID orderId) {
+        final DeliveryOrder order = deliveryOrderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+
+        final OrderStatus status = order.getStatus();
+
+        if (status != OrderStatus.DELIVERED) {
+            throw new IllegalStateException();
+        }
+
+        order.setStatus(OrderStatus.COMPLETED);
+        return order;
+    }
+
+    @Transactional(readOnly = true)
+    public List<DeliveryOrder> findAll() {
+        return deliveryOrderRepository.findAll();
+    }
+
+}
