+++ src/main/java/kitchenpos/order/deliveryorder/external/domain/KitchenridersClient.java
@@ -1,0 +1,8 @@
+package kitchenpos.order.deliveryorder.external.domain;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+public interface KitchenridersClient {
+    void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
+}
