+++ src/main/java/kitchenpos/order/eatinorder/domain/repository/EatInOrderRepository.java
@@ -1,0 +1,21 @@
+package kitchenpos.order.eatinorder.domain.repository;
+
+import kitchenpos.order.common.model.OrderStatus;
+import kitchenpos.order.common.model.OrderTable;
+import kitchenpos.order.eatinorder.domain.model.EatInOrder;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface EatInOrderRepository {
+
+    EatInOrder save(EatInOrder order);
+
+    Optional<EatInOrder> findById(UUID id);
+
+    List<EatInOrder> findAll();
+
+    boolean existsByOrderTableAndStatusNot(OrderTable orderTable, OrderStatus status);
+
+}
