+++ src/main/java/kitchenpos/order/eatinorder/domain/repository/OrderTableRepository.java
@@ -1,0 +1,16 @@
+package kitchenpos.order.eatinorder.domain.repository;
+
+import kitchenpos.order.common.model.OrderTable;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface OrderTableRepository {
+    OrderTable save(OrderTable orderTable);
+
+    Optional<OrderTable> findById(UUID id);
+
+    List<OrderTable> findAll();
+}
+
