+++ src/main/java/kitchenpos/order/eatinorder/infra/repository/JpaEatInOrderRepository.java
@@ -1,0 +1,10 @@
+package kitchenpos.order.eatinorder.infra.repository;
+
+import kitchenpos.order.eatinorder.domain.model.EatInOrder;
+import kitchenpos.order.eatinorder.domain.repository.EatInOrderRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaEatInOrderRepository extends EatInOrderRepository, JpaRepository<EatInOrder, UUID> {
+}
