+++ src/main/java/kitchenpos/order/eatinorder/representation/EatInOrderRestController.java
@@ -1,0 +1,47 @@
+package kitchenpos.order.eatinorder.representation;
+
+import kitchenpos.order.eatinorder.application.EatInOrderService;
+import kitchenpos.order.eatinorder.domain.model.EatInOrder;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.List;
+import java.util.UUID;
+
+@RequestMapping("/api/eat-in-orders")
+@RestController
+public class EatInOrderRestController {
+    private final EatInOrderService orderService;
+
+    public EatInOrderRestController(final EatInOrderService orderService) {
+        this.orderService = orderService;
+    }
+
+    @PostMapping
+    public ResponseEntity<EatInOrder> create(@RequestBody final EatInOrder request) {
+        final EatInOrder response = orderService.create(request);
+        return ResponseEntity.created(URI.create("/api/eat-in-orders/" + response.getId()))
+                .body(response);
+    }
+
+    @PutMapping("/{orderId}/accept")
+    public ResponseEntity<EatInOrder> accept(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(orderService.accept(orderId));
+    }
+
+    @PutMapping("/{orderId}/serve")
+    public ResponseEntity<EatInOrder> serve(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(orderService.serve(orderId));
+    }
+
+    @PutMapping("/{orderId}/complete")
+    public ResponseEntity<EatInOrder> complete(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(orderService.complete(orderId));
+    }
+
+    @GetMapping
+    public ResponseEntity<List<EatInOrder>> findAll() {
+        return ResponseEntity.ok(orderService.findAll());
+    }
+}
