--- src/main/java/kitchenpos/order/infra/repository/JpaOrderRepository.java
@@ -1,10 +1,0 @@
-package kitchenpos.order.infra.repository;
-
-import kitchenpos.order.domain.model.Order;
-import kitchenpos.order.domain.repository.OrderRepository;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface JpaOrderRepository extends OrderRepository, JpaRepository<Order, UUID> {
-}
