--- src/main/java/kitchenpos/order/infra/repository/JpaOrderTableRepository.java
@@ -1,10 +1,0 @@
-package kitchenpos.order.infra.repository;
-
-import kitchenpos.order.domain.model.OrderTable;
-import kitchenpos.order.domain.repository.OrderTableRepository;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
-}
