+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -3,9 +3,10 @@ package kitchenpos.order.application;
 import kitchenpos.external.infra.FakeKitchenridersClient;
 import kitchenpos.menu.domain.repository.MenuRepository;
 import kitchenpos.menu.infra.repository.InMemoryMenuRepository;
-import kitchenpos.order.domain.model.*;
-import kitchenpos.order.domain.repository.OrderRepository;
-import kitchenpos.order.domain.repository.OrderTableRepository;
+import kitchenpos.order.common.application.OrderService;
+import kitchenpos.order.common.domain.repository.OrderRepository;
+import kitchenpos.order.common.model.*;
+import kitchenpos.order.eatinorder.domain.repository.OrderTableRepository;
 import kitchenpos.order.infra.repository.InMemoryOrderRepository;
 import kitchenpos.order.infra.repository.InMemoryOrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
